pyproject.toml
[project]
description = "Python package to generate information cards from your Wuthering Wave's convene history."
name = "wuwaconvene"
requires-python = ">=3.10"

[project.urls]
homepage = "https://github.com/DEViantUA/StarRailCard"

[tool.poetry]
authors = ["DeviantUa <deviantapi@gmail.com>", "seria <seria.ati@gmail.com>"]
description = "Python package to generate information cards from your Wuthering Wave's convene history."
license = "GPL-3.0"
name = "wuwaconvene"
packages = [{include = "wuwaconvene"}]
readme = "README.md"
version = "0.1.0"

[tool.poetry.dependencies]
aiofiles = "^23.2.1"
aiohttp = "^3.9.5"
anyio = "^4.4.0"
beautifulsoup4 = "^4.12.3"
cachetools = "^5.3.3"
imageio = "^2.34.1"
more-itertools = "^10.2.0"
moviepy = "^1.0.3"
numpy = "^1.26.4"
pillow = "^10.3.0"
pydantic = "^2.7.3"
python = "^3.10"
kuro = {git = "https://github.com/Wuthery/kuro.py"}

[tool.poetry.group.dev]
optional = true

[tool.poetry.group.dev.dependencies]
pre-commit = "^3.7.1"
ruff = "^0.4.5"

[build-system]
build-backend = "poetry.core.masonry.api"
requires = ["poetry-core"]
